backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.20",
    "langchain-openai>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "python-dotenv>=1.1.1",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "ipykernel>=6.30.1",
    "pypdf>=6.0.0",
    "beautifulsoup4>=4.13.4",
    "arxiv>=2.2.0",
    "faiss-cpu>=1.12.0",
    "sentence-transformers>=5.1.0",
    "pandas>=2.3.2",
    "openai>=1.101.0",
    "langchain-groq>=0.3.7",
]
